requirements.txt
PyPDF2>=3.0.0
transformers>=4.30.0
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
spacy>=3.5.0
nltk>=3.8.1
datasets>=2.12.0
tqdm>=4.65.0
python-dotenv>=1.0.0
langchain>=0.0.300
sentence-transformers>=2.2.2
PyMuPDF>=1.21.1
Pillow>=9.5.0
timm>=0.9.2
torchvision>=0.15.0
optuna>=3.0.0
wandb>=0.15.0
pytorch-lightning>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.13.0
jinja2>=3.1.2
streamlit>=1.30.0
lime>=0.2.0
shap>=0.44.0
protobuf>=4.25.0 